Main Page Fragments Use Event Service

+++ app/src/main/java/com/example/eventhunter/ui/reservationDetailsCard/reservationCardPopup/ReservationCardDialogModel.java
@@ -13,9 +13,9 @@ public class ReservationCardDialogModel {
     public List<CollaboratorHeader> collaborators;
     public int availableSeatsNumber;
     public int chosenSeatsNumber;
-    public int ticketPrice;
+    public Double ticketPrice;
 
-    public ReservationCardDialogModel(String eventId, String eventName, String organizerName, String eventDate, String eventLocation, int availableSeatsNumber, int ticketPrice, List<CollaboratorHeader> collaborators) {
+    public ReservationCardDialogModel(String eventId, String eventName, String organizerName, String eventDate, String eventLocation, int availableSeatsNumber, Double ticketPrice, List<CollaboratorHeader> collaborators) {
         this.eventId = eventId;
         this.eventName = eventName;
         this.organizerName = organizerName;
